mount /boot rw from /dev/mmcblk0p1
This commit is the result of a headache problem when trying to replace
the uImage file in /boot: the /boot directory content was from the
rootfs but not from the root partition, meaning that any change in this
directory would not affect the uImage used after reboot.

To fix this problem, this commit provides:

- Automatic rw mount of /dev/mmcblk0p1 on /boot
- Deletion of /boot/* files from the rootfs to just keep the mount point

The sync option has been proposed for /boot mount point in order to be
sure that kernel image override are not cached in case of power outage
to avoid partially written image.

This commit was tested on NanoPi Neo SBC.
Here is a sample output after startup:

	root@nanopi-neo:~# ls /boot
	boot.scr  image-version-info  sun8i-h3-nanopi-neo.dtb  uImage

	root@nanopi-neo:~# mount
	/dev/mmcblk0p2 on / type ext4 (rw,relatime)
	devtmpfs on /dev type devtmpfs (rw,relatime,...)
	proc on /proc type proc (rw,relatime)
	sysfs on /sys type sysfs (rw,relatime)
	tmpfs on /run type tmpfs (rw,nosuid,nodev,mode=755)
	tmpfs on /var/volatile type tmpfs (rw,relatime)
	/dev/mmcblk0p1 on /boot type vfat (rw,relatime,sync,...)
	devpts on /dev/pts type devpts (rw,relatime,...)

Signed-off-by: Romain Porte <microjoe@microjoe.org>